Fix client side functionality of ecosystem page
